test: move FLRW-specific test
